Stop a tightened limit from trapping the storage already inside it
Five fixes from review, four of which an operator could actually hit:

An administrator lowering the per-vault maximum (or cutting an account's quota)
below what a vault already holds locked everyone out of their own storage: every
reclaim still ended above the new bound, so the checks refused it and the only exit
was deleting the vault. Both ceilings now bound INCREASES only — moving toward
compliance is always allowed, moving further away never is.

`dockvault.py storage --set-gb 1.5` wrote a fractional value into a setting typed as
an int, so the deployment it then recreated refused its own configuration and would
not start. The setting is a float now (sizing a small deployment at 1.5 GB is a
reasonable thing to ask for), and the .env value is formatted without exponent
notation or a stray '.0' so what is written is always what can be read back.

`--set-gb 0` reported success while meaning the opposite: 0 in this variable is "no
ceiling", the same as -1. It is refused now, and says where a real stop lives — the
live limit in the vault's own Settings, whose 0 does mean zero.

Two concurrent repairs of the same vault's ledger both inserted the owner's row and
the loser hit the uniqueness constraint as a 500; that path is self-healing, so it
re-reads the winner's row instead.

The contribution box displayed a rounded GB figure but saved the rounding, so
pressing Save without editing rewrote the allocation — and on a full vault the
rounded-down value was rejected outright. An untouched field now sends back the
exact byte figure it was rendered from.

Also drops a helper and an import left unused by the earlier commit.

+++ app/core/config.py
@@ -149,7 +149,9 @@ class Settings(BaseSettings):
     # this number stays in control of the maximum. Only bytes actually written count toward it:
     # empty vaults, however many, cost nothing. The earlier PLAN_MAX_STORAGE_GB spelling is still
     # accepted so an existing deployment's .env keeps working across an update.
-    max_storage_gb: int = Field(
+    # A float, not an int: an operator sizing a small deployment reasonably writes 1.5, and an
+    # int field would refuse the whole settings model and stop the deployment from starting.
+    max_storage_gb: float = Field(
         default=-1,
         validation_alias=AliasChoices("max_storage_gb", "plan_max_storage_gb"),
     )
